+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## New Commands
 - `/levelup`: Check exp for character and skills and level up the default character accordingly.
+- `/end_encounter`: GM utility to grant pips at the end of an encounter.
+- `/damage`: GM utility to create a button allowing players to take assigned damage.
+- `/heal`: Heal yourself for a given amount, or create a button for receiving healing.
 
 ## Bug Fixes
 - Fixed urls not being correctly parsed if they contained a `?`